fix(oauth2): Fixes API calls using OAuth2.
Gate calls were subtly broken due to OAuth2 redirects.
GET calls were traversing the redirects fine, but POST
calls would be turned into GETs with no body, which
the Gate Controllers fail on. Now, when we resolve the
oauth2 token we immediately establish a session for the life
of the spin command so all calls are authenticated.