chore(dev): Add go.mod and go.sum.
This program didn't track dependencies before, so I'm not at all sure
what it intended to do about its dependencies. In lieu of that, I'm
just going to add Go modules, because it's the easiest.

Updates the README to include instructions for both GOPATH and Go
modules development, in the event that organizations do not currently
approve use of Go versions with modules support.