feat(gateclient): Add Google service account auth config.
- Add support for using Google service accounts for oauth2.
- Add a configuration struct for Google service accounts.

This allows a user to either load a service account from a file or use
application default credentials for oauth2. In order for this to work
for a user, the user must run `gcloud auth application-default login`,
set the GOOGLE_APPLICATION_CREDENTIALS environment variable, or set the
`file` field of the `google_service_account` auth config.

Due to the maybe unusual layout of auth login calls, the
authenticateGoogleServiceAccount call happens after the
authenticateOAuth2 call.

The YAML writing part has been relocated to a writeYAML function
because both the oauth2 and Google service account login methods cache
tokens.