test(jsonpath): add tests for non-primitives
This change adds tests to prevent unintentional breakage related to
output of non-primitive jsonpath outputs.

The changes made in https://github.com/spinnaker/spin/pull/241 do not
pass these tests, as they return Go-syntax.

Updating client-go to v0.19.0-rc.2 that contains the fix for printing
JSON still does not pass the tests as the output would be encapsulated
in `[]`