feat(spin): Run container as non-root user and exec spin
Following the principle of least privilege we should not run as root
user. There is no need.
Match other Spinnaker services and run as a high user id.

Additional containers generally run the application process as
the entrypoint rather than a shell.
This makes performing commands simpler as we can tack on any arguments
to the end of the image like so:

```
$ docker run --rm -it spin:latest --version
version dev
```

The `VERSION` build arg now defaults to `dev` otherwise an empty VERSION
string overwrites default in Go and results in the `--version` flag
disappearing.