feat(ci): GHA - tag master built container as ":latest"
We have a convention in Spinnaker land of tagging containers with
`master-latest-unvalidated`.
`master` comes from the branch (master|release-*) and for `master` could
potentially be dropped.
`unvalidated` is used because of nightly integration tests which haven't
been used with `spin` before.

All in all push `:latest` tag on successful CI run for master commits.
Standard UX in container-land.